осталось доделать функцию подсчета упоминаний в статьях
