+++ src/about.html
@@ -5,8 +5,7 @@
   <meta name = "description" content="Сервис для анализа происходящих в мире событий">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <link rel="stylesheet" href="./pages/index.css">
-    <link rel="stylesheet" href="<%=htmlWebpackPlugin.files.chunks.about.css %>">
+  <link rel="stylesheet" href="<%=htmlWebpackPlugin.files.chunks.about.css %>">
   <link rel="icon" href="<%=require('./images/Favicon.png').default%>">
   <title>О проекте</title>
 </head>
@@ -89,7 +88,7 @@
           </div>
           <!--НАУЧИТСЯ ДЕЛАТЬ ЭТОТ БЛОК СЛАЙДЕР  -->
           <div class="carousel">
-<!-- пример карточки
+          <!-- пример карточки
             <div class="carousel-cell">
               <div class="position">
                 <time class="carusel__data">14 апреля 2019</time>
@@ -103,7 +102,7 @@
                 <p class="carusel__text">Emmet (formerly Zen Coding) is a web-developer’s toolkit that can greatly improve your HTML & CSS workflow.</p>
               </div>
             </div>
-          -->
+
             <ol class="flickity-page-dots">
               <li class="dot" aria-label="Page dot 1"></li>
               <li class="dot" aria-label="Page dot 2"></li>
@@ -111,6 +110,7 @@
               <li class="dot" aria-label="Page dot 4"></li>
               <li class="dot" aria-label="Page dot 5"></li>
             </ol>
+            -->
           </div>
           <a href="https://github.com/Psihosomatika/popularNews/commits/master" class="history__link_btn link" >Открыть в GitHub</a>
         </section>
