+++ src/analitica.html
@@ -6,7 +6,6 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <link rel="stylesheet" href="<%=htmlWebpackPlugin.files.chunks.analitica.css %>">
-  <link rel="stylesheet" href="./pages/index.css">
   <link rel="icon" href="<%=require('./images/Favicon.png').default%>">
   <title>О проекте</title>
 </head>
@@ -34,7 +33,7 @@
         <section class="analitica">
           <p class="section"><span class="section_black">Главная</span>/ Аналитика</p>
           <div class="analitica__reply">
-            <h2 class="analitica__title">Вы спросили: <span class="keyword"></span></h2><!--«Путешествия»-->
+            <h2 class="analitica__title">Вы спросили: <span class="keyword">nhfnfnf</span></h2><!--«Путешествия»-->
             <p class="analitica__text">Новостей за неделю: <span class="number-of-news-items bold"></span></p>
             <p class="analitica__text">Упоминаний в заголовках: <span class="mentions bold"></span></p>
           </div>
@@ -101,8 +100,7 @@
         </section>
       </footer>
   </div>
-  <script type="module" src="./js/analitica.js"></script>
-  <!--<script type="module" src="<%= htmlWebpackPlugin.files.chunks.analitica.entry %>"></script>-->
+  <script type="module" src="<%= htmlWebpackPlugin.files.chunks.analitica.entry %>"></script>
 </body>
 </html>
