+++ src/index.html
@@ -5,11 +5,8 @@
   <meta name = "description" content="Сервис для анализа происходящих в мире событий">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <link rel="stylesheet" href="./pages/index.css">
-
- <!-- <link rel="stylesheet" href="<%=htmlWebpackPlugin.files.chunks.main.css %>">-->
- <!--<link rel="icon" href="<%=require('./images/Favicon.png').default%>" type="image/ico">-->
-  <link rel="icon" href="./images/Favicon.png" type="image/ico">
+  <link rel="stylesheet" href="<%=htmlWebpackPlugin.files.chunks.main.css %>">
+  <link rel="icon" href="<%=require('./images/Favicon.png').default%>" type="image/ico">
   <title>Сервис для анализа происходящих в мире событий</title>
 </head>
 <body>
@@ -74,11 +71,10 @@
             <p class="circle-preloader__text">идет поиск</p>
           </div>
         </div>
-        <!--НИЧЕГО НЕ НАЙДЕН-->
+        <!--НИЧЕГО НЕ НАЙДЕНО-->
         <div class="cards__nothing-found">
           <div class="nothing-found__img">
-            <!--<img  src="<%=require('./images/not-found_v1.png').default%>" alt="Грустный смайлик">-->
-            <img  src="./images/not-found_v1.png" alt="Грустный смайлик">
+            <img  src="<%=require('./images/not-found_v1.png').default%>" alt="Грустный смайлик">
           </div>
           <p class="nothing-found__mini-text">Ничего не найдено</p>
           <p class="nothing-found__text">К сожалению по вашему запросу ничего не найдено.</p>
@@ -89,8 +85,7 @@
       <div class="cover__about-author">
         <section class="about-author">
           <div class="about-author__photo-me">
-            <img  class="photo" alt="фотография автора сайта" src="./images/image-03.png">
-            <!--<img  class="photo" alt="фотография автора сайта" src="<%=require('./images/image-03.png').default%>">-->
+            <img  class="photo" alt="фотография автора сайта" src="<%=require('./images/image-03.png').default%>">
           </div>
           <div>
             <h2 class="about-author__title">Об авторе</h2>
@@ -119,7 +114,6 @@
       </section>
     </footer>
   </div>
-  <script src="./js/index.js"></script>
-  <!--<script src="<%= htmlWebpackPlugin.files.chunks.main.entry %>"></script>-->
+  <script src="<%= htmlWebpackPlugin.files.chunks.main.entry %>"></script>
 </body>
 </html>
