+++ src/js/components/commitCardList.js
@@ -1,0 +1,9 @@
+export default class CommitCardList {
+  constructor(list) {
+    this.list = list;
+    this.renderCards = this.renderCards.bind(this);
+  }
+  renderCards(caruselCell) {
+    this.list.append(caruselCell);
+  }
+}
