+++ src/js/modules/githubApi.js
@@ -1,12 +1,11 @@
-
-class GithubApi {
-  constructor (url) {
+export default class GithubApi {
+  constructor(url) {
     this.url = url;
   }
-  getCommits (){
-    return fetch (url)
-
-    .then(res => res.ok ? Promise.resolve(res.json()) : Promise.reject(`Ошибка: ${res.status}`))
+  async getCommits() {
+    const res = await fetch(this.url);
+    return await (res.ok ? Promise.resolve(res.json()) : Promise.reject(`Ошибка: ${res.status}`));
   }
 }
+
