+++ src/js/modules/newsApi.js
@@ -1,6 +1,6 @@
-import {URL_START, URL_END, MS_IN_DAY} from '../constants/constants.js'
+import { MS_IN_DAY} from '../constants/constants.js'
 
-class NewsApi {
+export default class NewsApi {
   constructor (baseUrl, parameters) {
     this.baseUrl = baseUrl;
     this.parameters = parameters;
@@ -13,10 +13,9 @@ class NewsApi {
     const previousDate = new Date(currentDate.getTime()-previousDateInMs);
     const fromDate = previousDate.toISOString().slice(0, 10);
     const toDate = currentDate.toISOString().slice(0,10);
-    console.log(`${this.baseUrl}${searchWord}&from=${fromDate}&to=${toDate}&language=ru&${this.parameters}`);
 
     const res = await fetch(`${this.baseUrl}${searchWord}&from=${fromDate}&to=${toDate}&language=ru&${this.parameters}`);
     return await (res.ok ? Promise.resolve(res.json()) : Promise.reject(`Ошибка: ${res.status}`));
   }
 }
-const newsApi = new NewsApi(URL_START, URL_END);
+
